requirements.txt
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.3
requests==2.31.0
beautifulsoup4==4.12.3 